Move reporting assertions to reporting.testing
